Move the Binder stuff into its own module.
