+++ src/Judgement.hs
@@ -3,7 +3,6 @@ module Judgement where
 
 import Context hiding (S)
 import Data.Functor.Classes
-import Data.Functor.Foldable hiding (Nil)
 import Expr
 import Module
 import Prelude hiding (fail)
@@ -30,40 +29,6 @@ data Judgement a where
 
   Normalize :: Expr -> Judgement Expr
   WHNF :: Expr -> Judgement Expr
-
-
-class Binder a where
-  (<?) :: Name -> a -> Bool
-
-class Binder1 f where
-  liftIn :: (Name -> a -> Bool) -> Name -> f a -> Bool
-
-instance (Foldable t, Binder a) => Binder (t a) where
-  (<?) name = any (name <?)
-
-instance Binder Name where
-  (<?) = (==)
-
-instance Binder Binding where
-  name <? (_ := m) = name <? m
-
-instance Binder TermEntry where
-  name <? (_ ::: s) = name <? s
-
-instance Binder1 f => Binder (Fix f) where
-   (<?) name = liftIn (<?) name . unfix
-
-instance Binder1 (ExprF Name) where
-  liftIn occurs name expr = case expr of
-    Abs n _ | n == name -> False
-    Var v | v == name -> True
-    _ -> any (occurs name) expr
-
-instance Binder Entry where
-  n <? t = case t of
-    Ty d -> n <? d
-    Tm d -> n <? d
-    Sep -> False
 
 
 -- Instances
