+++ surface.cabal
@@ -23,6 +23,7 @@ library
                      , Module
                      , Parser
                      , REPL
+                     , Surface.Binder
                      , Surface.Proof
                      , Text.Pretty
   build-depends:       base >= 4.7 && < 5
